+++ Packages/eu.netherlands3d.selection-tools/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.8.0] - 2025-06-02
+
+### Added
+
+* Made area selection events public
+* Added area select reselect and clear function
+
 ## [2.7.1] - 2025-04-01
 
 ### Fixed